requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
torch
numpy
pytest
